[refactor] 폴더구조 수정
