+++ src/components/DetailBody.js
@@ -1,7 +1,7 @@
-import { React, useState, useEffect, useRef, useMemo } from 'react';
+import { React, useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import axios from 'axios';
-import Pagination from './Pagination';
+import Pagination from './footer/Pagination';
 import styled from 'styled-components';
 
 const ProductImg = styled.img`
